requirements.txt
absl-py==1.3.0
albumentations==1.3.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
attrs @ file:///croot/attrs_1668696182826/work
Babel @ file:///tmp/build/80754af9/babel_1620871417480/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
beautifulsoup4 @ file:///opt/conda/conda-bld/beautifulsoup4_1650462163268/work
bleach==5.0.1
brotlipy @ file:///home/conda/feedstock_root/build_artifacts/brotlipy_1648854164153/work
bunch==1.0.1
cachetools==5.2.0
certifi @ file:///croot/certifi_1665076670883/work/certifi
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1625835301720/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1661170624537/work
click==8.1.3
contextlib2==21.6.0
cryptography @ file:///home/conda/feedstock_root/build_artifacts/cryptography_1652967085355/work
cycler==0.11.0
debugpy==1.6.3
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
depthwise-conv2d-implicit-gemm==0.0.0
docker-pycreds==0.4.0
dropblock==0.3.0
einops==0.6.0
entrypoints @ file:///tmp/build/80754af9/entrypoints_1649908585034/work
fastjsonschema @ file:///opt/conda/conda-bld/python-fastjsonschema_1661371079312/work
filelock==3.8.0
fonttools==4.38.0
gitdb==4.0.9
GitPython==3.1.29
google-auth==2.14.1
google-auth-oauthlib==0.4.6
grpcio==1.50.0
h5py==3.7.0
huggingface-hub==0.11.1
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1663625384323/work
imageio==2.22.4
importlib-metadata==5.0.0
importlib-resources==5.10.0
ipykernel==6.16.2
ipython==7.34.0
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
ipywidgets==8.0.2
jedi==0.17.2
Jinja2 @ file:///croot/jinja2_1666908132255/work
joblib==1.2.0
json5 @ file:///tmp/build/80754af9/json5_1624432770122/work
jsonschema==4.17.1
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-lsp==1.5.1
jupyter-server==1.23.3
jupyter_client @ file:///croot/jupyter_client_1669040270237/work
jupyter_core @ file:///croot/jupyter_core_1668084443243/work
jupyterlab @ file:///tmp/abs_12f3h01vmy/croots/recipe/jupyterlab_1658907535764/work
jupyterlab-language-pack-zh-CN==3.5.post1
jupyterlab-lsp==3.10.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
jupyterlab_server @ file:///opt/conda/conda-bld/jupyterlab_server_1664908231281/work
kiwisolver==1.4.4
lab==7.2
loguru==0.6.0
lxml @ file:///opt/conda/conda-bld/lxml_1657545139709/work
Markdown==3.4.1
MarkupSafe @ file:///opt/conda/conda-bld/markupsafe_1654597864307/work
matplotlib==3.5.3
matplotlib-inline @ file:///opt/conda/conda-bld/matplotlib-inline_1662014470464/work
mistune==2.0.4
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
ml-collections==0.1.1
nbclassic @ file:///croot/nbclassic_1668174957779/work
nbclient==0.7.0
nbconvert==7.2.5
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==2.6.3
notebook @ file:///croot/notebook_1668179881751/work
notebook_shim @ file:///croot/notebook-shim_1668160579331/work
numpy @ file:///opt/conda/conda-bld/numpy_and_numpy_base_1653915516269/work
oauthlib==3.2.2
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1602866521163/work
opencv-contrib-python==4.7.0.72
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas==1.3.5
pandocfilters @ file:///opt/conda/conda-bld/pandocfilters_1643405455980/work
parso==0.7.1
pathtools==0.1.2
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.3.0
pkgutil_resolve_name @ file:///opt/conda/conda-bld/pkgutil-resolve-name_1661463321198/work
prometheus-client==0.15.0
promise==2.3
prompt-toolkit==3.0.33
protobuf==3.20.1
psutil==5.9.4
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
Pygments==2.13.0
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1663846997386/work
pyparsing @ file:///opt/conda/conda-bld/pyparsing_1661452539315/work
pyrsistent==0.19.2
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1648857264451/work
pysodmetrics==1.3.1
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz @ file:///opt/conda/conda-bld/pytz_1654762638606/work
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
qudida==0.0.4
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1661872987712/work
requests-oauthlib==1.3.1
rsa==4.9
ruamel.base==1.0.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
segmentation==0.2.2
Send2Trash @ file:///tmp/build/80754af9/send2trash_1632406701022/work
sentry-sdk==1.11.1
setproctitle==1.3.2
shortuuid==1.0.11
simplejson==3.18.0
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sklearn==0.0.post1
smmap==5.0.0
sniffio==1.3.0
soupsieve @ file:///croot/soupsieve_1666296392845/work
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
terminado==0.17.0
threadpoolctl==3.1.0
tifffile==2021.11.2
timm==0.6.12
tinycss2 @ file:///croot/tinycss2_1668168815555/work
torch==1.12.0
torch-tb-profiler==0.4.1
torchaudio==0.12.0
torchmetrics==0.11.4
torchstat==0.0.7
torchsummary==1.5.1
torchvision==0.13.0
tornado @ file:///opt/conda/conda-bld/tornado_1662061693373/work
tqdm==4.64.1
traitlets==5.5.0
ttach==0.0.3
txt2tags==3.7
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1665144421445/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1658789158161/work
wandb==0.13.5
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.2
wget==3.2
widgetsnbextension==4.0.3
xeus-python==0.14.3
xeus-python-shell==0.5.0
yacs==0.1.8
zipp==3.10.0
